Remove is_small(...) and use slightly different assertion.
This function was broken: in the original paper by Dekker (1971) as well
as Joldes et al. (2018), they claim that the condition for x.add_small(y)
to work is that the exponents of x and y must be ordered.  But in the
follow-up paper Muller et al (2022), they relaxed this condition, namely
that one is allowed to "denormalize" x, thereby increasing its exponent,
as long as one this not truncate the mantissa.  Thanks to the authors
for pointing this out to me.

This is much more challenging to check against, so we omit it. Instead,
we do a bare-bones check that what this yields is a valid double-double.